sitesense/requirements.txt
fastapi==0.115.12
httpx==0.28.1
protobuf==6.31.1
pydantic==2.11.5
python-dotenv==1.1.0

aiohttp

uvicorn
sse-starlette
google-genai
jinja2
bcrypt
